Add files via upload
https://medium.com/@ankan.ab21/terrorism-database-dashboard-5e5b990c2295

INSIGHTS FROM THE DASHBOARD:

The dashboard includes visualizations of the following:

a. Number of terrorist attacks over time

b. Attack locations on a map

c. Trends in attack types and weapons used

d. Comparisons of casualties by region or year

The dashboard also includes slicers and filters, which allow users to explore the data in different ways. For example, users can filter the data by region, time period, attack type, or weapon used.

<iframe title=”Report Section” width=”600" height=”373.5" src=”https://app.powerbi.com/view?r=eyJrIjoiNmU4YmYxNTQtODM4OS00ZmE2LTk1YmQtNGVjMGFlM2JiYTdjIiwidCI6ImRmODY3OWNkLWE4MGUtNDVkOC05OWFjLWM4M2VkN2ZmOTVhMCJ9" frameborder=”0" allowFullScreen=”true”></iframe>

💡 KEY FINDINGS:

The following are some key findings from the dashboard:

✅ The number of terrorist attacks has declined in recent years, but the global death toll from terrorism remains high.

✅The Middle East and North Africa (MENA) region is the most affected by terrorism, followed by South Asia and Sub-Saharan Africa.

✅The most common types of terrorist attacks are bombings, shootings, and armed assaults.

✅Suicide bombings and explosives are the deadliest weapons used in terrorist attacks.

✅The vast majority of victims of terrorism are civilians.

✔ RECOMMENDATIONS:

There is a need to address the root causes of terrorism and to develop effective counter-terrorism measures. Some specific recommendations include:

✅ Promoting peace and development in conflict-prone areas. This can help to reduce the risk of radicalization and recruitment into terrorist groups.

✅Strengthening international cooperation on counter-terrorism. This includes sharing intelligence, disrupting terrorist networks, and preventing the flow of foreign fighters.

✅Countering terrorist propaganda and ideology. This can help to reduce the appeal of terrorism and to build resilience against radicalization.

✅Protecting vulnerable populations. This includes providing security for religious and ethnic minorities and addressing the needs of victims of terrorism.

✳ In addition to these specific recommendations, it is also important to remember that terrorism is a complex problem with no easy solutions. It will require a sustained effort from all stakeholders, including governments, civil society, and the private sector, to address this threat.

Conclusion

This Terrorism Dashboard is my first task as a #powerBI developer intern at InternCareer. The goal of this project is to create a comprehensive and informative dashboard that visualizes global terrorism trends and patterns.